Fixed a critical bug, and in the process of fixing another

+++ src/main.ts
@@ -30,20 +30,34 @@ async function waitForBackendReady(retries = 20, delayMs = 250): Promise<void> {
 	throw new Error("Backend server did not become ready in time.");
 }
 async function cleanup_backend() {
-	await fetch("http://localhost:8000/download/serialize", { method: "POST", headers: { ...DefaultRequestOpts.headers, "x-session-token": uuidOfBackend! }, body: JSON.stringify({ filepath_to_serialize_to: serialize_path } as SerializeRequest) })
-	await fetch("http://localhost:8000/shutdown", { method: "POST", headers: { ...DefaultRequestOpts.headers, "x-session-token": uuidOfBackend! } })
-	console.log("Shut down the server")
-	if (backendProcess?.pid) {
-		console.log(`Killing backend PID: ${backendProcess.pid}`);
-		treeKill(backendProcess.pid, (err) => {
-			if (err) console.error('treeKill error:', err);
-			else console.log('treeKill success');
-		});
+	if (backendProcess === null) {
+		return
 	}
-	else {
-		backendProcess?.kill()
+	try {
+		console.log("Starting shutdown")
+		await fetch("http://localhost:8000/download/serialize", { method: "POST", headers: { ...DefaultRequestOpts.headers, "x-session-token": uuidOfBackend! }, body: JSON.stringify({ filepath_to_serialize_to: serialize_path } as SerializeRequest) })
+		console.log("Serialization complete")
+		await fetch("http://localhost:8000/shutdown", { method: "POST", headers: { ...DefaultRequestOpts.headers, "x-session-token": uuidOfBackend! } })
+		console.log("Shut down the server")
 	}
-	backendProcess = null;
+	catch (err) {
+		console.error(err)
+	}
+	finally {
+
+		if (backendProcess?.pid) {
+			console.log(`Killing backend PID: ${backendProcess.pid}`);
+			treeKill(backendProcess.pid, (err) => {
+				if (err) console.error('treeKill error:', err);
+				else console.log('treeKill success');
+			});
+		}
+		else {
+			backendProcess?.kill()
+		}
+		backendProcess = null;
+		console.log("Done shutting down!")
+	}
 }
 
 
@@ -119,6 +133,7 @@ app.on('activate', () => {
 
 
 app.on("before-quit", cleanup_backend)
+app.on("will-quit", cleanup_backend)
 
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and import them here.
